Add clean deletion and seeding feature

+++ bundles/Javra/RojanBundle/Helper/DeleteHelper.php
@@ -5,6 +5,7 @@ namespace Javra\RojanBundle\Helper;
 use Pimcore\Model\Asset;
 use Pimcore\Model\DataObject;
 use Pimcore\Model\DataObject\Book;
+use Pimcore\Model\DataObject\Category;
 
 class DeleteHelper{
     private string $class_name;
@@ -14,29 +15,50 @@ class DeleteHelper{
     }
     
     public function deleteAssets(){
+        //exclude deletion of root node
+        $root_node_paths =  ['/', '/Book'];
         $assets = new Asset\Listing();
-        $assets->setCondition("filename like 'book asset %'");
+        $assets->load();
         foreach($assets as $asset)
-        {
-            echo $asset->getFileName();
-            echo "<br>";
-            $asset->delete();
+        { 
+            if( ! in_array($asset->getFullPath(), $root_node_paths) ){
+                echo $asset->getFileName(). " deleted";
+                echo "<br>";
+                $asset->delete();
+            }
         }
         echo "Asset deleted successfully using " .$this->class_name. "<br>";
     }
 
     public function deleteDataObjects(){
-        $objs = new DataObject\Book\Listing();
-        $objs->setUnpublished(true);
-        $objs->setCondition("title like 'Book%'");
-        $objs->load();
+        self::deleteBookObjects();
+        self::deleteCategoryObjects();
+    }
+
+    private function deleteBookObjects(){
+       //delete book
+       $obj_list = new DataObject\Book\Listing();
+       $obj_list->setUnpublished(true);
+       // $objs->setCondition("title like 'Book%'");
+       $obj_list->load();
+       self::deleteObjects($obj_list);
+       echo "Book objects deletion success!! <br>";
+    }
+
+    private function deleteCategoryObjects(){
+        $obj_list = new DataObject\Category\Listing();
+        $obj_list->load();
+        self::deleteObjects($obj_list);
+        echo "Category objects deletion success!! <br>";
+
+    }
+
+    private function deleteObjects($objs){
         foreach($objs as $obj)
         {   
-            echo $obj->getTitle();
+            echo $obj->getKey();
             echo "<br>";
             $obj->delete();
         }
-        echo "object deleted successfully using " .$this->class_name. "<br>";
     }
-
 }
